baseline/requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
autopep8==1.6.0
bert4keras==0.10.9
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.3
cycler==0.10.0
dataclasses==0.8
Django==3.2.12
et-xmlfile==1.1.0
filelock==3.3.1
flatbuffers==1.12
gast==0.3.3
gensim==4.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.43.0
h5py==2.10.0
huggingface-hub==0.0.19
idna==3.2
importlib-metadata==4.6.4
jieba==0.42.1
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
nltk==3.6.5
numpy==1.19.5
oauthlib==3.1.1
openpyxl @ file:///tmp/build/80754af9/openpyxl_1632777717936/work
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
Pillow==8.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pynvml==11.4.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.4.1
six==1.15.0
smart-open==5.2.1
sqlparse==0.4.2
tensorboard==2.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
